aruco manipulation stack cleanup done
